+++ test/OpenTelemetry.Sampler.AWS.Tests/TestAWSXRayRemoteSampler.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Diagnostics;
+using System.Reflection;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
 using WireMock.RequestBuilders;
@@ -18,27 +19,34 @@ public class TestAWSXRayRemoteSampler
     {
         var pollingInterval = TimeSpan.FromSeconds(5);
         var endpoint = "http://localhost:3000";
-
-        var sampler = AWSXRayRemoteSampler.Builder(ResourceBuilder.CreateEmpty().Build())
+        var parentBasedSampler = AWSXRayRemoteSampler.Builder(ResourceBuilder.CreateEmpty().Build())
             .SetPollingInterval(pollingInterval)
             .SetEndpoint(endpoint)
-            .BuildXraySampler();
+            .Build();
 
-        Assert.Equal(pollingInterval, sampler.PollingInterval);
-        Assert.Equal(endpoint, sampler.Endpoint);
-        Assert.NotNull(sampler.RulePollerTimer);
-        Assert.NotNull(sampler.Client);
+        FieldInfo? rootSamplerFieldInfo = typeof(ParentBasedSampler).GetField("rootSampler", BindingFlags.NonPublic | BindingFlags.Instance);
+
+        var xraySampler = (AWSXRayRemoteSampler?)rootSamplerFieldInfo?.GetValue(parentBasedSampler);
+
+        Assert.Equal(pollingInterval, xraySampler?.PollingInterval);
+        Assert.Equal(endpoint, xraySampler?.Endpoint);
+        Assert.NotNull(xraySampler?.RulePollerTimer);
+        Assert.NotNull(xraySampler?.Client);
     }
 
     [Fact]
     public void TestSamplerWithDefaults()
     {
-        var sampler = AWSXRayRemoteSampler.Builder(ResourceBuilder.CreateEmpty().Build()).BuildXraySampler();
+        var parentBasedSampler = AWSXRayRemoteSampler.Builder(ResourceBuilder.CreateEmpty().Build()).Build();
 
-        Assert.Equal(TimeSpan.FromMinutes(5), sampler.PollingInterval);
-        Assert.Equal("http://localhost:2000", sampler.Endpoint);
-        Assert.NotNull(sampler.RulePollerTimer);
-        Assert.NotNull(sampler.Client);
+        FieldInfo? rootSamplerFieldInfo = typeof(ParentBasedSampler).GetField("rootSampler", BindingFlags.NonPublic | BindingFlags.Instance);
+
+        var xraySampler = (AWSXRayRemoteSampler?)rootSamplerFieldInfo?.GetValue(parentBasedSampler);
+
+        Assert.Equal(TimeSpan.FromMinutes(5), xraySampler?.PollingInterval);
+        Assert.Equal("http://localhost:2000", xraySampler?.Endpoint);
+        Assert.NotNull(xraySampler?.RulePollerTimer);
+        Assert.NotNull(xraySampler?.Client);
     }
 
     [Fact(Skip = "Flaky test. Related issue: https://github.com/open-telemetry/opentelemetry-dotnet-contrib/issues/1219")]
@@ -53,7 +61,7 @@ public class TestAWSXRayRemoteSampler
             .SetPollingInterval(TimeSpan.FromMilliseconds(10))
             .SetEndpoint(mockServer.Url!)
             .SetClock(clock)
-            .BuildXraySampler();
+            .Build();
 
         // the sampler will use fallback sampler until rules are fetched.
         Assert.Equal(SamplingDecision.RecordAndSample, this.DoSample(sampler, "cat-service"));